requirements.txt
# Базовые зависимости для работы с PostgreSQL и парсинга
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Для работы с данными
pandas==2.1.4
numpy==1.26.2

# Для парсинга веб-данных (УК/ТСЖ)
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Для веб-интерфейса (будет позже)
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-multipart==0.0.6

# Для работы с Excel (если понадобится)
openpyxl==3.1.2
xlrd==2.0.1

# Утилиты
tqdm==4.66.1
python-dateutil==2.8.2
